+++ tests/all_tests.rs
@@ -1,4 +1,6 @@
 #![deny(unused_mut)]
+#![cfg_attr(feature = "adt_const_marker", feature(adt_const_params))]
+#![cfg_attr(feature = "adt_const_marker", allow(incomplete_features))]
 #![cfg_attr(feature = "nightly_mut_refs", feature(const_mut_refs))]
 
 mod misc_tests {